lastupdates

+++ src/app/admin/missions/missions.component.ts
@@ -2,6 +2,7 @@ import { HttpErrorResponse } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { UsersService } from 'src/app/services/users.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-missions',
@@ -10,46 +11,47 @@ import { UsersService } from 'src/app/services/users.service';
 })
 export class MissionsComponent implements OnInit {
 
-  p:number = 1 ;
-  dataStudent={
-    id : '',
-    title:'',
-    description:''
-   /* averagePayment:0 ,
-    period:0,
-    start_date:'',
-    end_date:'',
-    */
+  p: number = 1;
+  dataStudent = {
+    id: '',
+    title: '',
+    description: ''
+    /* averagePayment:0 ,
+     period:0,
+     start_date:'',
+     end_date:'',
+     */
   }
-  messageErr =''
-  
-  messageSuccess = '' ;
-  title: string ="" ;
-  searchedKeyword!:string;
+  messageErr = ''
+
+  messageSuccess = '';
+  title: string = "";
+  searchedKeyword!: string;
   dataArray: any;
 
-  constructor(private usersService:UsersService,private route:Router) {
-    
+  constructor(private usersService: UsersService, private route: Router) {
+
   }
-  messageErrr =''
+  messageErrr = ''
 
 
   ngOnInit(): void {
-    this.usersService.getAllMissions().subscribe(data=>{
-      console.log(data)
-      this.dataArray=data , (err:HttpErrorResponse)=>{
-        console.log(err)
-      this.messageErr="We dont't found this student in our database"} 
-      //console.log(this.dataArray)
-    }) 
+    this.usersService.getAllMissions().subscribe(data => {
+
+      this.dataArray = data, (err: HttpErrorResponse) => {
+
+        this.messageErr = "We dont't found this student in our database"
+      }
+
+    })
 
     /*
     this.produitServiceService.getAllusers().subscribe(data=>{
-      console.log(data)
+
       this.dataArrays=data , (err:HttpErrorResponse)=>{
-        console.log(err)
+ 
       this.messageErrr="We dont't found this user in our database"} 
-      //console.log(this.dataArray)
+    
     }) 
     */
   }
@@ -65,54 +67,65 @@ export class MissionsComponent implements OnInit {
   }
   */
 
-  key = 'id' ;
-  reverse: boolean = false ;
+  key = 'id';
+  reverse: boolean = false;
 
   sort(key: string) {
-    this.key = key ;
-    this.reverse = !this.reverse ;
+    this.key = key;
+    this.reverse = !this.reverse;
   }
-  details(id:any){
-    this.route.navigate(['/posts/'+id])
+  details(id: any) {
+    this.route.navigate(['/posts/' + id])
   }
 
 
-  delete(id:any  , i :number){
 
-    this.usersService.deleteMission(id).subscribe(response=>{
-      console.log(response)
-      this.dataArray.splice(i,1)
+  delete(id: any, i: number) {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.usersService.deleteMission(id).subscribe(response => {
+          this.dataArray.splice(i, 1)
+        })
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+    })
+
+
+  }
+  getdata(title: string, description: string, id: any) {
+    this.messageSuccess = ''
+    this.dataStudent.title = title
+    this.dataStudent.description = description
+    this.dataStudent.id = id
+
+  }
+  updatenewstudent(f: any) {
+    let data = f.value
+    this.usersService.updateMission(this.dataStudent.id, data).subscribe(response => {
+      let indexId = this.dataArray.findIndex((obj: any) => obj.id == this.dataStudent.id)
+
+      this.dataArray[indexId].title = data.title
+      this.dataArray[indexId].description = data.description
+
+      this.messageSuccess = `this title : ${this.dataArray[indexId].title} is updated`
+
+    }, (err: HttpErrorResponse) => {
 
     })
-    
   }
-  
-    getdata(title:string,description:string,id:any){
-      this.messageSuccess=''
-      this.dataStudent.title=title
-      this.dataStudent.description=description
-      this.dataStudent.id=id
-      console.log(this.dataStudent)
-  
-    }
-    updatenewstudent(f:any){
-      let data=f.value
-      this.usersService.updateMission(this.dataStudent.id,data).subscribe(response=>
-        {
-        console.log(response)
-          let indexId=this.dataArray.findIndex((obj:any)=>obj.id==this.dataStudent.id)
-  
-          this.dataArray[indexId].title=data.title
-          this.dataArray[indexId].description=data.description
-  
-          this.messageSuccess=`this title : ${this.dataArray[indexId].title} is updated`
-  
-        },(err:HttpErrorResponse)=>{
-          console.log(err.message)
-        
-        })
-    }
-  
-  
+
+
 
 }
